+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [[v0.2.0]](https://github.com/mlange-42/arche-pixel/compare/v0.1.0...v0.2.0)
+
+### Features
+
+* New drawer `Inspector` for inspecting entities (#21)
+
 ## [[v0.1.0]](https://github.com/mlange-42/arche-pixel/compare/v0.0.3...v0.1.0)
 
 ### Features